Nearly complete
I am uploading a version of the PowerPoint that is nearly complete. All that is missing are the images from the program that are not committed to GitHub yet. I will be uploading an updated version of the slides write up as well. I will also add more to one of the PowerPoint slides after I complete the write up to add more information.